100 prueba4

+++ pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@ int _atoi(char *s)
 	int i;
 	int multiplier = 10;
 	int negatigncount = 0;
-	long num = 0;
+	int num = 0;
 
 	for (i = 0; s[i] != '\0'; i++)
 	{
